Plurals fix for other languages
Now when we add new strings to the newres translation set, we
add the plurals for the target language. Also, when generating
the pseudo translated text, we use the right plurals for the
target language.

The end result is that the new strings file has the right set of
plurals in it now, and pseudo localized files also have the
right set of plurals.