Make this able to read already-localized po files
The problem was that the default template did not apply to all
po files so localized po files were not getting a locale when
they were read in. The output file name template is used to
construct a regular expression to recognize already localized
files and what the locale of that file is. Without the template,
the locale was never extracted and the source and target were
both en-US which is not correct.