Make PO files into resource file types
... that is they can be the destination file type for the resources
extracted from other file types like Python or C++ with gettext.
They can also be used as the destination for conversions. Resources
that are added to the PO file for a particular locale and which
do not have translations for that locale or which have translations
for a different locale will be treated as source-only resources.
The idea is that the string will exist in the resource file as a
placeholder, but will not contain translations.