Fix incorrectly localized url resources
Fixed a bug where URLs in direct links were added to the new strings set
even when the localize-links directive was turned off. If the "fully
translated" flag was also turned on, then the plugin would think a file was
not fully translated because these links appear in the new strings set
and so it would not produce the localized version of the file.