Add support for new replacements and tags
In parsing the resources, you can now extracted the
formatted, the context, and the flavor attributes for a resource.

In creating a plural template, you can now use those three
attributes.

When getting the value of an attribute, you can now use the
_pathname and _basename values, which will convert into the full
path name of the file, and the base name of the file.