Change tactics. Put the context into the reskey
The problem is that the loctool and other tools do not differentiate
properly between two resources that have the same key but different
contexts. When they are added together to the same translation set,
they override one another instead of both existing in the set. When
the context is put into the resource key, then they are
differentiated properly.