Fix a problem loading data from the cache
- if the loader does not support sync loading and sync loading is
requested, then loadData() should check the cache. If the data is
in the cache already (by previous loads or by ensureLocale) then
the data can be returned synchronously anyways.
- that way, you can do ensureLocale() at the top of your program
and always load locale data synchronously after that
- this mechanism was somewhat broken before as we didn't check
the cache and throw an exception if the data was not there
- also fixed DataCache.clearCache() so that it resets the cache
instead of creating a whole new one. The problem was if you have
an existing pointer to the old cache through a LocaleData object,
and then call LocaleData.clearCache() you get a whole new cache
object, but the LocaleData instance did not. It is supposed to be
a singleton. There should never be 2 instances of it!