Disassemble and the reassemble the locale data
That way, the resource file data can mix with the ilib locale
data and then get output to the same files. The locale data
from the modules is added to the superset by each sublocale.
The resource data is added by its locale. Then, on output, each
locale is decomposed into an array of sublocales, and the
merged data for all sublocales is added to the output file.