Fixed infinite recursion bug
Indented HTML that was the only thing on the line was not parsed
correctly, causing it to attempt to reparse recursively over and
over again. This was fixed by using the whitespace trimmed version
of the tag to ignore the indenting and parse the tag syntax
correctly.

Added a unit test to show the error and that it was fixed.