Fix unsafe Resource metadata access
`Resource` is not a `Record<string, string>` but a proper class with
a specific interface. Attempting to access a property using a generic `string` key
is unsafe and can lead to runtime errors. This change ensures that:
1. we only access properties that are known to exist on the `Resource` class
2. we only set properties that have `string` values
3. we only directly set properties that allow it
4. we use the `set` method for those properties that do not allow direct setting
5. we throw a meaningful error if the metadata fix command attempts to perform an operation
   that is not allowed on the `Resource` class (e.g. setting a property that does not exist)

To achieve this, it's necessary to explicitly list out all supported fields
within the command's `apply` method. In future, (once we migrate to TS)
we could try looking into extracting a `const string` union type for the supported settable fields,
but for now, this is the best way to avoid undefined behaviour or runtime errors.