Convert fix and command creation methods to static
These methods do not rely on anything specific from within
of the fixer instance. Additionally, in order to use them
one must already know that they are using specifically the `ResourceFixer`
(as those methods are not available in the abstract `Fixer` class).
On the flip side, making those methods static means rules
don't need to instantiate the fixer anymore, which simplifies the code.