resourceFixer review changes (#98)
* Fix incorrect docs on MetadataFixCommand

* Bugfix invalid property access

`getName` and `getValue` methods don't exist on the `ResourceMetadataFixCommand`.

* Remove empty constructor

It's unnecessary to provide an empty constructor.
See https://eslint.org/docs/latest/rules/no-useless-constructor

* Remove unused applied flag on command

It should only be possible to apply the whole fix - never just a single command.
Keeping this flag around is unnecessary and may cause confusion.

* Fix unsafe Resource metadata access

`Resource` is not a `Record<string, string>` but a proper class with
a specific interface. Attempting to access a property using a generic `string` key
is unsafe and can lead to runtime errors. This change ensures that:
1. we only access properties that are known to exist on the `Resource` class
2. we only set properties that have `string` values
3. we only directly set properties that allow it
4. we use the `set` method for those properties that do not allow direct setting
5. we throw a meaningful error if the metadata fix command attempts to perform an operation
   that is not allowed on the `Resource` class (e.g. setting a property that does not exist)

To achieve this, it's necessary to explicitly list out all supported fields
within the command's `apply` method. In future, (once we migrate to TS)
we could try looking into extracting a `const string` union type for the supported settable fields,
but for now, this is the best way to avoid undefined behaviour or runtime errors.

* Remove unnecessary return value

This flag is only used to throw an error outside of the `apply` method (and also never set to `false`),
so it's less confusing to just throw the error immediately.

* Remove redundant getters on ResourceFix

Instead of introducing boilerplate getters, we can safely expose the properties directly
because they are `readonly`. Getters and setters should only be declared when
there's an explicit need for them i.e. there's something special in what they do
(e.g. when some setter also trips the dirty flag, or when a getter has to pull its value
from some nested property).

* Remove unused import

* Clean up undeclared property

`ResourceStringFixCommand#stringFix` property was not declared in the class.
Added an explicit property declaration and removed the now-redundant getter.

* Clean up shadowed variables and improve readability in ResourceFixer

Slightly refactored the ResourceFixer loop to prevent accidentally
marking wrong fixes as applied.

* Remove redundant getters on ResourceStringLocator

* Remove unused return value

The "success" return value was never used. Additionally, I think this should probably
throw instead of silently failing - but we can think about that later.

* Convert fix and command creation methods to static

These methods do not rely on anything specific from within
of the fixer instance. Additionally, in order to use them
one must already know that they are using specifically the `ResourceFixer`
(as those methods are not available in the abstract `Fixer` class).
On the flip side, making those methods static means rules
don't need to instantiate the fixer anymore, which simplifies the code.

* [Tentative] Expose available resource fix commands as "verbs" to be used when constructing a fix (on par with StringFixCommand)

- Simplified ResourceStringFixCommand to just be a container for the underlying StringFixCommand - there are no additional properties it holds because Locator is a property of the Fix, so its only purpose is to introduce a common abstract class for available ResourceFixCommands; removed redundant arg checks for string fix commands - this is handled by StringFixCommand class already
- Re-exported StringFixCommand static "verbs" on ResourceStringFixCommand
- Added `set` static "verb" on metadata command class for parity with string command
- Collected available commands under static `ResourceFixer.commands` - having them attached like this allows for easier access (i.e. without additional imports), which I imagine was the original purpose for exposing `ResourceFixer.createMetadataCommand` and `ResourceFixer.createStringCommand`

* Revert "[Tentative] Expose available resource fix commands as "verbs" to be used when constructing a fix (on par with StringFixCommand)"

This reverts commit 9cf06bcc9474a3ffac34687ba52428a861dd96ee.